Move 'NoteFsService' to shared folder
Took 5 minutes
