+++ src/app/note/finder/finder.component.spec.ts
@@ -18,9 +18,9 @@ import { NoteCalendarComponent } from '../calendar/calendar.component';
 import { NoteMetadataDummyFactory } from '../dummies';
 import { NoteItemComponent } from '../item/item.component';
 import { NoteMetadata } from '../models';
-import { NoteFsService } from '../note-fs.service';
 import { noteReducerMap, NoteStateWithRoot } from '../reducers';
 import { NoteCollectionSortingMenu } from '../shared/note-collection-sorting.menu';
+import { NoteFsService } from '../shared/note-fs.service';
 import { NoteFinderComponent } from './finder.component';
