+++ src/app/note/shared/note-fs.service.spec.ts
@@ -1,10 +1,10 @@
 import { fakeAsync, flush, inject, TestBed } from '@angular/core/testing';
 import * as path from 'path';
 import { of } from 'rxjs/observable/of';
-import { MockFsService } from '../../testing/mock';
-import { FsService } from '../core/fs.service';
-import { NoteContentDummyFactory, NoteMetadataDummyFactory } from './dummies';
-import { NoteContent, NoteMetadata } from './models';
+import { MockFsService } from '../../../testing/mock';
+import { FsService } from '../../core/fs.service';
+import { NoteContentDummyFactory, NoteMetadataDummyFactory } from '../dummies';
+import { NoteContent, NoteMetadata } from '../models';
 import { NoteFsService } from './note-fs.service';
